Exported funcition VerifyData
