+++ index.html
@@ -10,8 +10,7 @@
 		<link rel="shortcut icon" href="./images/iconeaba.jpg" >
 		<title>GEMA-Descubra seus dados de forma fácil</title>
 		<script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-		<script src="./Backend/Locate.js"></script>
-		<script src="./Backend/Search.js"></script>
+		<script src="./Backend/Locate.js" type="module"></script>
 		<script src="./Backend/Copy.js"></script>
 		<link rel="stylesheet" href="./styles/styles.css">
 		<link rel="stylesheet" href="./styles/media-queries.css">
